+++ pkg/server/cloudprovider/aws.go
@@ -3,6 +3,7 @@ package cloudprovider
 import (
 	"strconv"
 
+	"github.com/luizalabs/teresa/pkg/server/service"
 	"github.com/luizalabs/teresa/pkg/server/teresa_errors"
 )
 
@@ -27,4 +28,29 @@ func (ops *awsOperations) CreateOrUpdateSSL(appName, cert string, port int) erro
 		return teresa_errors.NewInternalServerError(err)
 	}
 	return nil
+}
+
+func (ops *awsOperations) SSLInfo(appName string) (*service.SSLInfo, error) {
+	var port int
+	an, err := ops.k8s.ServiceAnnotations(appName, appName)
+	if err != nil {
+		if ops.k8s.IsNotFound(err) {
+			return nil, ErrServiceNotFound
+		}
+		return nil, teresa_errors.NewInternalServerError(err)
+	}
+	sslAn, ok := an[awsSSLPortsAnnotation]
+	if ok {
+		port, err = strconv.Atoi(sslAn)
+		if err != nil {
+			return nil, teresa_errors.NewInternalServerError(err)
+		}
+	}
+	info := &service.SSLInfo{
+		Cert: an[awsSSLCertAnnotation],
+		ServicePort: &service.ServicePort{
+			Port: port,
+		},
+	}
+	return info, nil
 }