+++ pkg/server/cloudprovider/aws_test.go
@@ -24,3 +24,21 @@ func TestCreateOrUpdateSSLFail(t *testing.T) {
 		t.Errorf("got %v; want %v", teresa_errors.Get(err), e)
 	}
 }
+
+func TestSSLInfoSuccess(t *testing.T) {
+	ops := &awsOperations{&FakeK8sOperations{}}
+
+	if _, err := ops.SSLInfo("teresa"); err != nil {
+		t.Errorf("got %v; want no error", err)
+	}
+}
+
+func TestSSLInfoFail(t *testing.T) {
+	k8s := &FakeK8sOperations{ServiceAnnotationsErr: errors.New("test")}
+	ops := &awsOperations{k8s}
+
+	e := teresa_errors.ErrInternalServerError
+	if _, err := ops.SSLInfo("teresa"); teresa_errors.Get(err) != e {
+		t.Errorf("got %v; want %v", teresa_errors.Get(err), e)
+	}
+}